refactor (#427)
